kindle_screen_shot/requirements.txt
pyautogui
pywinauto
Pillow
PyQt6
pyinstaller